+++ automation/src/epilot/models/shared/activitytrigger.py
@@ -5,7 +5,7 @@ import dataclasses
 from dataclasses_json import Undefined, dataclass_json
 from enum import Enum
 from epilot import utils
-from typing import Optional
+from typing import Final, Optional
 
 class ActivityTriggerConfigurationTypes(str, Enum):
     CREATE_METER_READING = 'CreateMeterReading'
@@ -23,15 +23,12 @@ class ActivityTriggerConfiguration:
     
 
 
-class ActivityTriggerType(str, Enum):
-    ACTIVITY = 'activity'
-
 
 @dataclass_json(undefined=Undefined.EXCLUDE)
 
 @dataclasses.dataclass
 class ActivityTrigger:
     configuration: ActivityTriggerConfiguration = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('configuration') }})
-    type: ActivityTriggerType = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})
+    TYPE: Final[str] = dataclasses.field(default='activity', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('type') }})