+++ automation/src/epilot/models/shared/cartcheckoutconfig.py
@@ -23,7 +23,7 @@ class CartCheckoutConfigMappingAttributes:
 
 @dataclasses.dataclass
 class CartCheckoutConfig:
-    linkback_relation_attribute: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('linkback_relation_attribute'), 'exclude': lambda f: f is None }})
+    linkback_relation_attribute: Optional[str] = dataclasses.field(default='mapped_entities', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('linkback_relation_attribute'), 'exclude': lambda f: f is None }})
     r"""Relation attribute on the main entity where the target entity will be linked. Set to false to disable linkback"""
     linkback_relation_tags: Optional[list[str]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('linkback_relation_tags'), 'exclude': lambda f: f is None }})
     r"""Relation tags (labels) to include in main entity linkback relation attribute"""
@@ -32,7 +32,7 @@ class CartCheckoutConfig:
     relation_attributes: Optional[list[shared_relationattribute.RelationAttribute]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('relation_attributes'), 'exclude': lambda f: f is None }})
     target_unique: Optional[list[str]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('target_unique'), 'exclude': lambda f: f is None }})
     r"""Unique key for target entity (see upsertEntity of Entity API)"""
-    version: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('version'), 'exclude': lambda f: f is None }})
+    version: Optional[str] = dataclasses.field(default='1', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('version'), 'exclude': lambda f: f is None }})
     r"""Version of the config"""