+++ automation/src/epilot/models/shared/mapentityconfig.py
@@ -25,7 +25,7 @@ class MapEntityConfigMappingAttributes:
 class MapEntityConfig:
     target_schema: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('target_schema') }})
     r"""Schema of target entity"""
-    linkback_relation_attribute: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('linkback_relation_attribute'), 'exclude': lambda f: f is None }})
+    linkback_relation_attribute: Optional[str] = dataclasses.field(default='mapped_entities', metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('linkback_relation_attribute'), 'exclude': lambda f: f is None }})
     r"""Relation attribute on the main entity where the target entity will be linked. Set to false to disable linkback"""
     linkback_relation_tags: Optional[list[str]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('linkback_relation_tags'), 'exclude': lambda f: f is None }})
     r"""Relation tags (labels) to include in main entity linkback relation attribute"""